@@ -1,50 +1,51 @@
 {
-	"name": "ngx-flexi-table",
-	"version": "0.0.0",
-	"license": "MIT",
-	"scripts": {
-		"ng": "ng",
-		"start": "ng serve",
-		"build": "ng build --prod",
-		"test": "ng test",
-		"lint": "ng lint",
-		"e2e": "ng e2e",
-		"packagr": "ng-packagr -p ng-package.json"
-	},
-	"private": false,
-	"dependencies": {
-		"@angular/animations": "^5.2.0",
-		"@angular/common": "^5.2.0",
-		"@angular/compiler": "^5.2.0",
-		"@angular/core": "^5.2.0",
-		"@angular/forms": "^5.2.0",
-		"@angular/http": "^5.2.0",
-		"@angular/platform-browser": "^5.2.0",
-		"@angular/platform-browser-dynamic": "^5.2.0",
-		"@angular/router": "^5.2.0",
-		"core-js": "^2.4.1",
-		"rxjs": "^5.5.6",
-		"zone.js": "^0.8.19"
-	},
-	"devDependencies": {
-		"@angular/cli": "~1.7.1",
-		"@angular/compiler-cli": "^5.2.0",
-		"@angular/language-service": "^5.2.0",
-		"@types/jasmine": "~2.8.3",
-		"@types/jasminewd2": "~2.0.2",
-		"@types/node": "~6.0.60",
-		"codelyzer": "^4.0.1",
-		"jasmine-core": "~2.8.0",
-		"jasmine-spec-reporter": "~4.2.1",
-		"karma": "~2.0.0",
-		"karma-chrome-launcher": "~2.2.0",
-		"karma-coverage-istanbul-reporter": "^1.2.1",
-		"karma-jasmine": "~1.1.0",
-		"karma-jasmine-html-reporter": "^0.2.2",
-		"ng-packagr": "^2.1.0",
-		"protractor": "~5.1.2",
-		"ts-node": "~4.1.0",
-		"tslint": "^5.9.1",
-		"typescript": "^2.5.3"
-	}
+  "name": "ngx-flexi-table",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build --prod",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "packagr": "ng-packagr -p ng-package.json"
+  },
+  "private": false,
+  "dependencies": {
+    "@angular/animations": "6.0.3",
+    "@angular/common": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/forms": "6.0.3",
+    "@angular/http": "6.0.3",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
+    "@angular/router": "6.0.3",
+    "core-js": "^2.4.1",
+    "rxjs": "^6.2.0",
+    "zone.js": "^0.8.26"
+  },
+  "devDependencies": {
+    "@angular/cli": "^6.0.7",
+    "@angular/compiler-cli": "6.0.3",
+    "@angular/language-service": "6.0.3",
+    "@types/jasmine": "~2.8.3",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "~6.0.60",
+    "codelyzer": "^4.0.1",
+    "jasmine-core": "~2.8.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-packagr": "^2.1.0",
+    "protractor": "~5.1.2",
+    "ts-node": "~4.1.0",
+    "tslint": "^5.9.1",
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.6"
+  }
 }
